Struct webmachine_rust::context::WebmachineContext

pub struct WebmachineContext {
    pub request: WebmachineRequest,
    pub response: WebmachineResponse,
    pub selected_media_type: Option<String>,
    pub selected_language: Option<String>,
    pub selected_charset: Option<String>,
    pub selected_encoding: Option<String>,
    pub if_unmodified_since: Option<DateTime<FixedOffset>>,
    pub if_modified_since: Option<DateTime<FixedOffset>>,
    pub redirect: bool,
    pub new_resource: bool,
    pub metadata: HashMap<String, String>,
}

Main context struct that holds the request and response.

Fields

request

Request that the webmachine is executing against

response

Response that is the result of the execution

selected_media_type

selected media type after content negotiation

selected_language

selected language after content negotiation

selected_charset

selected charset after content negotiation

selected_encoding

selected encoding after content negotiation

if_unmodified_since

parsed date and time from the If-Unmodified-Since header

if_modified_since

parsed date and time from the If-Modified-Since header

redirect

If the response should be a redirect

new_resource

If a new resource was created

metadata

General store of metadata. You can use this to store attributes as the webmachine executes.

Methods

impl WebmachineContext

fn default() -> WebmachineContext

Creates a default context