Struct webmachine_rust::WebmachineResource

pub struct WebmachineResource {
    pub finalise_response: Option<Box<Fn(&mut WebmachineContext)>>,
    pub render_response: Box<Fn(&mut WebmachineContext) -> Option<String>>,
    pub available: Box<Fn(&mut WebmachineContext) -> bool>,
    pub known_methods: Vec<String>,
    pub uri_too_long: Box<Fn(&mut WebmachineContext) -> bool>,
    pub allowed_methods: Vec<String>,
    pub malformed_request: Box<Fn(&mut WebmachineContext) -> bool>,
    pub not_authorized: Box<Fn(&mut WebmachineContext) -> Option<String>>,
    pub forbidden: Box<Fn(&mut WebmachineContext) -> bool>,
    pub unsupported_content_headers: Box<Fn(&mut WebmachineContext) -> bool>,
    pub acceptable_content_types: Vec<String>,
    pub valid_entity_length: Box<Fn(&mut WebmachineContext) -> bool>,
    pub finish_request: Box<Fn(&mut WebmachineContext, &WebmachineResource)>,
    pub options: Box<Fn(&mut WebmachineContext, &WebmachineResource) -> Option<HashMap<String, Vec<String>>>>,
    pub produces: Vec<String>,
    pub languages_provided: Vec<String>,
    pub charsets_provided: Vec<String>,
    pub encodings_provided: Vec<String>,
    pub variances: Vec<String>,
    pub resource_exists: Box<Fn(&mut WebmachineContext) -> bool>,
    pub previously_existed: Box<Fn(&mut WebmachineContext) -> bool>,
    pub moved_permanently: Box<Fn(&mut WebmachineContext) -> Option<String>>,
    pub moved_temporarily: Box<Fn(&mut WebmachineContext) -> Option<String>>,
    pub is_conflict: Box<Fn(&mut WebmachineContext) -> bool>,
    pub allow_missing_post: Box<Fn(&mut WebmachineContext) -> bool>,
    pub generate_etag: Box<Fn(&mut WebmachineContext) -> Option<String>>,
    pub last_modified: Box<Fn(&mut WebmachineContext) -> Option<DateTime<FixedOffset>>>,
    pub delete_resource: Box<Fn(&mut WebmachineContext) -> Result<bool, u16>>,
    pub post_is_create: Box<Fn(&mut WebmachineContext) -> bool>,
    pub process_post: Box<Fn(&mut WebmachineContext) -> Result<bool, u16>>,
    pub create_path: Box<Fn(&mut WebmachineContext) -> Result<String, u16>>,
    pub process_put: Box<Fn(&mut WebmachineContext) -> Result<bool, u16>>,
    pub multiple_choices: Box<Fn(&mut WebmachineContext) -> bool>,
    pub expires: Box<Fn(&mut WebmachineContext) -> Option<DateTime<FixedOffset>>>,
}

Struct to represent a resource in webmachine

Fields

finalise_response

This is called just before the final response is constructed and sent. It allows the resource an opportunity to modify the response after the webmachine has executed.

render_response

This is invoked to render the response for the resource

available

Is the resource available? Returning false will result in a '503 Service Not Available' response. Defaults to true. If the resource is only temporarily not available, add a 'Retry-After' response header.

known_methods

HTTP methods that are known to the resource. Default includes all standard HTTP methods. One could override this to allow additional methods

uri_too_long

If the URI is too long to be processed, this should return true, which will result in a

allowed_methods

HTTP methods that are allowed on this resource. Defaults to GET','HEAD and 'OPTIONS'.

malformed_request

If the request is malformed, this should return true, which will result in a '400 Malformed Request' response. Defaults to false.

not_authorized

Is the client or request not authorized? Returning a Some will result in a '401 Unauthorized' response. Defaults to None. If a Some(String) is returned, the string will be used as the value in the WWW-Authenticate header.

forbidden

Is the request or client forbidden? Returning true will result in a '403 Forbidden' response. Defaults to false.

unsupported_content_headers

If the request includes any invalid Content-* headers, this should return true, which will result in a '501 Not Implemented' response. Defaults to false.

acceptable_content_types

The list of acceptable content types. Defaults to 'application/json'. If the content type of the request is not in this list, a '415 Unsupported Media Type' response is returned.

valid_entity_length

If the entity length on PUT or POST is invalid, this should return false, which will result in a '413 Request Entity Too Large' response. Defaults to true.

finish_request

This is called just before the final response is constructed and sent. This allows the response to be modified. The default implementation adds CORS headers to the response

options

If the OPTIONS method is supported and is used, this return a HashMap of headers that

produces

The list of content types that this resource produces. Defaults to 'application/json'. If more than one is provided, and the client does not supply an Accept header, the first one will be selected.

languages_provided

The list of content languages that this resource provides. Defaults to an empty list, which represents all languages. If more than one is provided, and the client does not supply an Accept-Language header, the first one will be selected.

charsets_provided

The list of charsets that this resource provides. Defaults to an empty list, which represents all charsets with ISO-8859-1 as the default. If more than one is provided, and the client does not supply an Accept-Charset header, the first one will be selected.

encodings_provided

The list of encodings your resource wants to provide. The encoding will be applied to the response body automatically by Webmachine. Default includes only the 'identity' encoding.

variances

The list of header names that should be included in the response's Vary header. The standard content negotiation headers (Accept, Accept-Encoding, Accept-Charset, Accept-Language) do not need to be specified here as Webmachine will add the correct elements of those automatically depending on resource behavior. Default is an empty list.

resource_exists

Does the resource exist? Returning a false value will result in a '404 Not Found' response unless it is a PUT or POST. Defaults to true.

previously_existed

If this resource is known to have existed previously, this should return true. Default is false.

moved_permanently

If this resource has moved to a new location permanently, this should return the new location as a String. Default is to return None

moved_temporarily

If this resource has moved to a new location temporarily, this should return the new location as a String. Default is to return None

is_conflict

If this returns true, the client will receive a '409 Conflict' response. This is only called for PUT requests. Default is false.

allow_missing_post

Return true if the resource accepts POST requests to nonexistent resources. Defaults to false.

generate_etag

If this returns a value, it will be used as the value of the ETag header and for comparison in conditional requests. Default is None.

last_modified

Returns the last modified date and time of the resource which will be added as the Last-Modified header in the response and used in negotiating conditional requests. Default is None

delete_resource

Called when a DELETE request should be enacted. Return Ok(true) if the deletion succeeded, and Ok(false) if the deletion was accepted but cannot yet be guaranteed to have finished. If the delete fails for any reason, return an Err with the status code you wish returned (a 500 status makes sense). Defaults to Ok(true).

post_is_create

If POST requests should be treated as a request to put content into a (potentially new) resource as opposed to a generic submission for processing, then this should return true. If it does return true, then create_path will be called and the rest of the request will be treated much like a PUT to the path returned by that call. Default is false.

process_post

If post_is_create returns false, then this will be called to process any POST request. If it succeeds, return Ok(true), Ok(false) otherwise. If it fails for any reason, return an Err with the status code you wish returned (e.g., a 500 status makes sense). Default is false. If you want the result of processing the POST to be a redirect, set context.redirect to true.

create_path

This will be called on a POST request if post_is_create returns true. It should create the new resource and return the path as a valid URI part following the dispatcher prefix. That path will replace the previous one in the return value of WebmachineRequest.request_path for all subsequent resource function calls in the course of this request and will be set as the value of the Location header of the response. If it fails for any reason, return an Err with the status code you wish returned (e.g., a 500 status makes sense). Default will return an Ok(WebmachineRequest.request_path). If you want the result of processing the POST to be a redirect, set context.redirect to true.

process_put

This will be called to process any PUT request. If it succeeds, return Ok(true), Ok(false) otherwise. If it fails for any reason, return an Err with the status code you wish returned (e.g., a 500 status makes sense). Default is Ok(true)

multiple_choices

If this returns true, then it is assumed that multiple representations of the response are possible and a single one cannot be automatically chosen, so a 300 Multiple Choices will be sent instead of a 200. Default is false.

expires

If the resource expires, this should return the date/time it expires. Default is None.

Methods

impl WebmachineResource

fn default() -> WebmachineResource

Creates a default webmachine resource